おかしいとこの修正を軽くやりました

+++ src/convert.py
@@ -62,14 +62,15 @@ def convert():
     for event in ret_dict:
         print(ret_dict[event])
     
-    with open('./dict_hashval', 'w') as fo:
-        print('[MIS] calculate hash from dict')
-        cat_summary = ""
-        for summary in ret_dict:
-            cat_summary.join(ret_dict[summary]['SUMMARY'])
-        dict_hash = hashlib.md5(cat_summary.encode()).hexdigest()
-        fo.write(dict_hash)
-        print(f"[MIS] write hash value, {dict_hash}")
+    if not os.path.exists('./dict_hashval'):
+        with open('./dict_hashval', 'w') as fo:
+            print('[MIS] calculate hash from dict')
+            cat_summary = ""
+            for summary in ret_dict:
+                cat_summary.join(ret_dict[summary]['SUMMARY'])
+            dict_hash = hashlib.md5(cat_summary.encode()).hexdigest()
+            fo.write(dict_hash)
+            print(f"[MIS] write hash value, {dict_hash}")
     with codecs.open('./calendar.json', 'w', 'utf-8') as fo:
         print('[MIS] write calendar.json')
         fo.write(json.dumps(ret_dict, indent=4, ensure_ascii=False))
